Revert "feat: Add cli test for GRD format COG"
